refactor smart explainer facade and move compute logic to Explainer
