+++ shapash/explainer/__init__.py
@@ -1,0 +1,6 @@
+"""Explainer package public exports."""
+
+from .explainer import Explainer
+from .smart_explainer import SmartExplainer
+
+__all__ = ["Explainer", "SmartExplainer"]
